# openapi/operioz.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Operioz REST API",
    "version": "0.1.0-poc",
    "description": "Façade REST du new-stack Operioz, exposée EN PARALLÈLE de tRPC (OPE-366).\nPoC : domaine clients. Cible : génération automatique depuis les schémas zod des routeurs tRPC via @fastify/swagger + fastify-zod-openapi. Auth : cookie `token` (JWT, credentials: include)."
  },
  "servers": [{ "url": "/", "description": "Edge public (staging.operioz.com) → dispatcher → new-stack" }],
  "tags": [{ "name": "clients", "description": "CRM — données personnelles, isolation tenant stricte" }],
  "paths": {
    "/api/rest/clients": {
      "get": {
        "operationId": "listClients",
        "tags": ["clients"],
        "summary": "Liste les clients du tenant courant",
        "responses": {
          "200": {
            "description": "Liste des clients (scopée tenant)",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Client" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NoArtisan" }
        }
      }
    },
    "/api/rest/clients/{id}": {
      "get": {
        "operationId": "getClientById",
        "tags": ["clients"],
        "summary": "Détail d'un client par identifiant",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Le client demandé",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Client" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Requête invalide",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Non authentifié (cookie token absent/invalide)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NoArtisan": {
        "description": "Utilisateur sans artisan (tenant non résolu)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Ressource introuvable (ou hors tenant — anti-oracle PII)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
      "ClientType": { "type": "string", "enum": ["particulier", "professionnel"] },
      "Client": {
        "type": "object",
        "required": ["id", "artisanId", "nom", "type", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "artisanId": { "type": "integer" },
          "nom": { "type": "string" },
          "prenom": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"] },
          "telephone": { "type": ["string", "null"] },
          "adresse": { "type": ["string", "null"] },
          "codePostal": { "type": ["string", "null"] },
          "ville": { "type": ["string", "null"] },
          "adresseFacturation": { "type": ["string", "null"] },
          "codePostalFacturation": { "type": ["string", "null"] },
          "villeFacturation": { "type": ["string", "null"] },
          "type": { "$ref": "#/components/schemas/ClientType" },
          "raisonSociale": { "type": ["string", "null"] },
          "siret": { "type": ["string", "null"] },
          "numeroTVA": { "type": ["string", "null"] },
          "etiquettes": { "type": ["string", "null"] },
          "notes": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}